Backend_Flask/requirements.txt
flask
pymongo
python-dotenv
sentence-transformers
scikit-learn
numpy
nltk
google-generativeai
